Split project from now defunct `commons-utilities`.
It did not make much sense to keep absolutely everything in one library,
so this project was split. As a result we have a much smaller .jar that
can be included when needed, without unnecessary bloat or unwanted
transitive dependencies.

This repo will have to be cleaned up as soon as another repo with collection
utilities is spawned and https://github.com/tguzik/annotations is avaliable
through Maven