Cleanup of the repository and added MutableValue<T>.
Removed most classes that fit more into a separate libraries and are not
directly related to value classes. Some classes (Normalize) recieved
new functionality (tabs to spaces) and additional unit tests.

MutableValue<T> is a wrapper on AtomicReference<T> that gives you access
to all goodies of normal, immutable value classes plus the ability to
change the reference after object has been instanitated.

NOTE: results of Value.toString() have changed - now only the toString()
of the underlying value is printed or, if the underlying value is null,
an empty string is printed.