Added PerformanceAwareBaseObject and JaxB java type adapter templates. Bumped version in pom.xml.
In PerformanceAwareBaseObject the implementor is forced to provide his own
implementations of hashCode() and equals() methods, so they won't rely
on reflection mechanisms.

JaxB adapters are just convenience classes really to make it easier to
work with value classes.

Bumped version to 0.3-SNAPSHOT to mark new functionality within the
project.